Use uint64 instead of int for task ID

+++ models/task.go
@@ -1,7 +1,7 @@
 package models
 
 type Task struct {
-	ID          int    `json:"id" gorm:"primary_key"`
+	ID          uint64 `json:"id" gorm:"primary_key"`
 	Title       string `json:"title"`
 	Description string `json:"description"`
 	Done        bool   `json:"done"`
@@ -20,8 +20,8 @@ type UpdateTaskTemplate struct {
 
 type TaskDB interface {
 	GetTasks() []*Task
-	GetTask(id int) (*Task, error)
+	GetTask(id uint64) (*Task, error)
 	CreateTask(t *NewTaskTemplate) *Task
 	UpdateTask(t *Task) error
-	DeleteTask(id int) error
+	DeleteTask(id uint64) error
 }
